Initial version
This initial version uses hardcoded values for its calculations so it only works for a comparison from 2004 - 2024. Some very basic groundwork has been done for using shortcode parameters to specify a different stock than Apple, and also start and end dates, if this was combined with some kind of stock lookup (not forgetting checking for any stock splits per company) then it could in theory be used to compare any stocks increase. But the initial aim of comparing 2004 to 2024 Apple was what I was trying to achieve with this initial release.